requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
# psycopg2-binary==2.9.9
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic== 1.10.22
email-validator==2.1.0
jinja2==3.1.2
aiosmtplib==3.0.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2







